Initial push of Main.cpp.
Main.cpp contains all of the code for the project (break it up into separate files later).

The "camera" is in a fixed position and fixed orientation, although the position can easily be modified. Orthogonal projection is used to render the objects, which means that two objects of the same size at different depths will appear as having the same size (the further object will not appear smaller, as it would with perspective projection).

The project allows for drawing triangle objects, where each vertex of a triangle can have a different depth. Triangles can intersect and will render appropriately.

Each triangle has an alpha value. If a non-opaque triangle is above other triangles (which may be opaque or partially transparent), then an appropriate mixture of pixel colors will display at that pixel location.